API Endpoint added except Quiz

+++ PRX/Controllers/Haghighi/HaghighiUserEducationStatusController.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
 using PRX.Data;
 using PRX.Dto.Haghighi;
 using PRX.Models.Haghighi;
@@ -28,16 +29,39 @@ namespace PRX.Controllers.Haghighi
 
         // GET: api/HaghighiUserEducationStatus/5
         [HttpGet("{id}")]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult GetHaghighiUserEducationStatusById(int id)
         {
-            var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.Id == id);
-            if (educationStatus == null)
+
+            try
             {
-                return NotFound();
+
+                // Retrieve the user ID from the token
+                var tokenUserId = int.Parse(User.FindFirst("id")?.Value);
+
+                // Ensure that the user is updating their own profile
+                if (id != tokenUserId)
+                {
+                    return Forbid(); // Or return 403 Forbidden
+                }
+                var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.UserId == id && !e.IsDeleted);
+                if (educationStatus == null)
+                {
+                    return NotFound();
+                }
+                return Ok(educationStatus);
+
             }
-            return Ok(educationStatus);
+
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Exception occurred: {ex}");
+                return BadRequest(new { Message = "Failed to update user profile." });
+            }
+
+
         }
 
         // POST: api/HaghighiUserEducationStatus
@@ -68,44 +92,88 @@ namespace PRX.Controllers.Haghighi
 
         // PUT: api/HaghighiUserEducationStatus/5
         [HttpPut("{id}")]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult UpdateHaghighiUserEducationStatus(int id, [FromBody] HaghighiUserEducationStatusDto educationStatusDto)
         {
-            var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.Id == id);
-            if (educationStatus == null)
+            try
             {
-                return NotFound();
+
+                // Retrieve the user ID from the token
+                var tokenUserId = int.Parse(User.FindFirst("id")?.Value);
+
+                // Ensure that the user is updating their own profile
+                if (id != tokenUserId)
+                {
+                    return Forbid(); // Or return 403 Forbidden
+                }
+                var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.UserId == id && !e.IsDeleted);
+                if (educationStatus == null)
+                {
+                    return NotFound();
+                }
+
+                educationStatus.UserId = educationStatusDto.UserId;
+                educationStatus.LastDegree = educationStatusDto.LastDegree;
+                educationStatus.FieldOfStudy = educationStatusDto.FieldOfStudy;
+                educationStatus.GraduationYear = educationStatusDto.GraduationYear;
+                educationStatus.IssuingAuthority = educationStatusDto.IssuingAuthority;
+
+                _context.SaveChanges();
+
+                return Ok(educationStatus);
+
             }
 
-            educationStatus.UserId = educationStatusDto.UserId;
-            educationStatus.LastDegree = educationStatusDto.LastDegree;
-            educationStatus.FieldOfStudy = educationStatusDto.FieldOfStudy;
-            educationStatus.GraduationYear = educationStatusDto.GraduationYear;
-            educationStatus.IssuingAuthority = educationStatusDto.IssuingAuthority;
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Exception occurred: {ex}");
+                return BadRequest(new { Message = "Failed to update user profile." });
+            }
 
-            _context.SaveChanges();
 
-            return Ok(educationStatus);
+   
         }
 
         // DELETE: api/HaghighiUserEducationStatus/5
         [HttpDelete("{id}")]
+        [Authorize]
         [ProducesResponseType(StatusCodes.Status200OK)]
         [ProducesResponseType(StatusCodes.Status404NotFound)]
         public IActionResult DeleteHaghighiUserEducationStatus(int id)
         {
-            var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.Id == id);
-            if (educationStatus == null)
+            try
             {
-                return NotFound();
+
+                // Retrieve the user ID from the token
+                var tokenUserId = int.Parse(User.FindFirst("id")?.Value);
+
+                // Ensure that the user is updating their own profile
+                if (id != tokenUserId)
+                {
+                    return Forbid(); // Or return 403 Forbidden
+                }
+                var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.UserId == id && !e.IsDeleted);
+                if (educationStatus == null)
+                {
+                    return NotFound();
+                }
+
+                _context.HaghighiUserEducationStatuses.Remove(educationStatus);
+                _context.SaveChanges();
+
+                return Ok();
+
             }
 
-            _context.HaghighiUserEducationStatuses.Remove(educationStatus);
-            _context.SaveChanges();
-
-            return Ok();
+            catch (Exception ex)
+            {
+                Console.WriteLine($"Exception occurred: {ex}");
+                return BadRequest(new { Message = "Failed to update user profile." });
+            }
+ 
         }
 
         // DELETE: api/HaghighiUserEducationStatus
@@ -116,6 +184,27 @@ namespace PRX.Controllers.Haghighi
             var educationStatuses = _context.HaghighiUserEducationStatuses.ToList();
             _context.HaghighiUserEducationStatuses.RemoveRange(educationStatuses);
             _context.SaveChanges();
+            return Ok();
+        }
+
+
+        // PUT: api/HaghighiUserProfile/complete/{id}
+        [HttpPut("complete/{id}")]
+        //[Authorize(Roles = "Admin")] // Assuming only admins can mark profiles as complete
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        public IActionResult MarkEducationStatusAsComplete(int id)
+        {
+            var educationStatus = _context.HaghighiUserEducationStatuses.FirstOrDefault(e => e.UserId == id);
+            if (educationStatus == null)
+            {
+                return NotFound();
+            }
+
+            educationStatus.IsComplete = true;
+            _context.SaveChanges();
+
             return Ok();
         }
     }
